+++ src/utils-constant/constant.js
@@ -1,3 +1,43 @@
+const TYPES = [
+  'Taxi',
+  'Bus',
+  'Train',
+  'Ship',
+  'Drive',
+  'Flight',
+  'Check-in',
+  'Sightseeing',
+  'Restaurant',
+];
+const CITIES = [
+  'New-York',
+  'Tokio',
+  'Saint-Petersburg',
+  'London',
+  'Paris',
+  'San-Francisco',
+  'Moscow',
+];
+const OFFERS = [
+  'Add luggage',
+  'Switch to comfort class',
+  'Add meal',
+  'Choose seats',
+  'Travel by train',
+  'Order Uber',
+  'Rent a car',
+  'Add breakfast',
+  'Book tickets',
+  'Lunch in city',
+];
+
+const EMPTY_LIST = {
+  'EVERYTHING': 'Click New Event to create your first point',
+  'LOADING': 'Loading...',
+  'LOADING_ERROR': 'Failed to load latest route information'
+};
+export { };
+
 const SortType = {
   DAY: 'Day',
   EVENT: 'Event',
@@ -6,4 +46,10 @@ const SortType = {
   OFFERS: 'Offers'
 };
 
-export { SortType };
+export {
+  SortType,
+  TYPES,
+  CITIES,
+  OFFERS,
+  EMPTY_LIST
+};
